# [POC][exporter] exporter batcher - byte size based batching