<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
## Description

This is an POC of serialized size based batching (version 2)

Configuration is supported via an additional field to `MaxSizeConfig`.
```
type MaxSizeConfig struct {
	MaxSizeItems int `mapstructure:"max_size_items"`
	MaxSizeBytes int `mapstructure:"max_size_bytes"`
}
```
We will validate that at most one of the above fields are specified (TODO) and switch between item count-based batching vs. byte size-based batching accordingly.

To get the byte size of otlp protos, this PR updates `pdata/internal/cmd/pdatagen/internal/templates/message.go.tmpl` to expose an interface `Size()`. This change will apply to all `pdatagen`-generated files.

-------------------------------------------------------------------------------------------------------------------------

## Getting serialized byte size

Getting serialized byte size is an expensive operation that gets repeatedly invoked during bytes based merge-splitting. The POC shows two optimizations that improves by 1. reducing size calculation times 2. reducing size calculation cost

### Caching size in `Request`

We repeated call `request.ByteSize()` while merg-splitting. By caching previous result in `request.ByteSize()`, we were able to improve the performance especially for the case where many requests are merged into a single batch.

```
BenchmarkSplittingBasedOnItemCountManySmallLogs-10                 	     324	   3729469 ns/op
/*Before*/ BenchmarkSplittingBasedOnByteSizeManySmallLogs-10         	       9	 120696384 ns/op
/*After*/ BenchmarkSplittingBasedOnByteSizeManySmallLogs-10                  358	   2906934 ns/op
```

### Proto size caculation

While filling the batch, we repeated evaluate size of new items and the remaining space, and both calculations are expensive. This PR shows an optimization that calculates remaining space iteratively using new item size:

```
remaning -= (1 + sizeof(newItem) + sov(sizeof(newItem)))
```

where

```
func sov(x uint64) (n int) {
	return (math_bits.Len64(x|1) + 6) / 7
}
```

This helped significantly improve the performance. For benchmark result without this optimization, see https://github.com/open-telemetry/opentelemetry-collector/pull/12017. 

## Performance Benchmark

### Benchmark 1 - 1000 small requests merge into one batch
This benchmark merges 1000 requests x 10 logs / request. The accumulated batch is ~1MB
```
BenchmarkSplittingBasedOnItemCountManySmallLogs-10                 	     328	   3556645 ns/op
BenchmarkSplittingBasedOnByteSizeManySmallLogs-10                  	     447	   2689311 ns/op
```


### Benchmark 2 - Every incoming request causes a split

**Case 1**: Merging 10 request, where each contains 10001 logs / ~ 1 MB and is slightly above the limit
```
BenchmarkSplittingBasedOnItemCountManyLogsSlightlyAboveLimit-10    	      45	  24271094 ns/op
BenchmarkSplittingBasedOnByteSizeManyLogsSlightlyAboveLimit-10     	      31	  35001137 ns/op
```

**Case 2**: Merging 10 request, where each contains 9999 logs / ~ 1 MB and is slightly below the limit
```
BenchmarkSplittingBasedOnItemCountManyLogsSlightlyBelowLimit-10    	      52	  21750268 ns/op
BenchmarkSplittingBasedOnByteSizeManyLogsSlightlyBelowLimit-10     	      38	  28904701 ns/op
```

### Benchmark 3 - A huge request splits into 10 batches
This benchmark merge splits a request with 100000 logs / ~10MB in to 10 batches.
```
BenchmarkSplittingBasedOnItemCountHugeLog-10                       	      43	  29004844 ns/op
BenchmarkSplittingBasedOnByteSizeHugeLog-10                        	      15	  66705669 ns/op
```




<!-- Issue number if applicable -->
#### Link to tracking issue
Fixes #3262

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->

TODO: ByteSize() should return int64 instead of int