@sfc-gh-sili 2025-01-16
TODO: Try if we can get away with not exposing that many size functions. Ideally we can do everything from one API.